# perf(rum): use instance fields over service factory abstraction[skip ci]