--- a/packages/rum-core/src/error-logging/index.js
+++ b/packages/rum-core/src/error-logging/index.js
@@ -25,20 +25,4 @@
 
 const ErrorLogging = require('./error-logging')
 
-module.exports = {
-  ErrorLogging,
-  registerServices: function registerServices(serviceFactory) {
-    serviceFactory.registerServiceCreator('ErrorLogging', function() {
-      const apmService = serviceFactory.getService('ApmServer')
-      const configService = serviceFactory.getService('ConfigService')
-      const loggingService = serviceFactory.getService('LoggingService')
-      const transactionService = serviceFactory.getService('TransactionService')
-      return new ErrorLogging(
-        apmService,
-        configService,
-        loggingService,
-        transactionService
-      )
-    })
-  }
-}
+module.exports = ErrorLogging