--- a/packages/rum-core/src/opentracing/index.js
+++ b/packages/rum-core/src/opentracing/index.js
@@ -26,21 +26,4 @@
 const Tracer = require('./tracer')
 const Span = require('./span')
 
-function createTracer(serviceFactory) {
-  var performanceMonitoring = serviceFactory.getService('PerformanceMonitoring')
-  var transactionService = serviceFactory.getService('TransactionService')
-  var errorLogging = serviceFactory.getService('ErrorLogging')
-  var loggingService = serviceFactory.getService('LoggingService')
-  return new Tracer(
-    performanceMonitoring,
-    transactionService,
-    loggingService,
-    errorLogging
-  )
-}
-
-module.exports = {
-  Span,
-  Tracer,
-  createTracer
-}
+module.exports = { Span, Tracer }