--- a/packages/rum/webpack.config.js
+++ b/packages/rum/webpack.config.js
@@ -33,7 +33,7 @@ const OUTPUT_DIR = path.join(__dirname, 'dist', 'bundles')
 const baseConfig = {
   entry: {
     'elastic-apm-rum': './src/index.js',
-    'elastic-apm-opentracing': './src/opentracing-entry.js'
+    'elastic-apm-opentracing': './src/opentracing/index.js'
   },
   output: {
     filename: '[name].umd.js',