… packages/rum-core/src/common/service-factory.js
… packages/rum-core/src/performance-monitoring/performance-monitoring.js
… packages/rum-core/test/common/service-factory.spec.js
… packages/rum/src/bootstrap.js
… packages/rum/src/index.js
… packages/rum/src/opentracing/index.js
… packages/rum/test/e2e/bundle-test/index.html