+ Part of elastic/apm-agent-rum-js#159 
+ Removed service factory pattern in favor of using singleton instances which is provided by the entry  point. 
+ Renamed service classes to be more appropriate (config, logger, stacktrace etc. )

Pending tasks
+ [ ] Add tests
+ [ ] Profile code to show before/after
+ [x] Remove the factory for Opentracing entry as well 

### Results

### Savings: Factory Pattern for services

Filename | Before(master) | After 
-------- | --------| ---------
elastic-apm-js-base.min.js |  54047  | **51570**
elastic-apm-opentracing.umd.min.js |  60432  | **57613**


### Initialization cost `elasticApm.init()`

**Before**
```
init-time: 3.22802734375ms
init-time: 3.3759765625ms
init-time: 3.239990234375ms
init-time: 2.166015625ms
init-time: 3.766845703125ms
init-time: 3.6962890625ms
init-time: 3.9951171875ms
init-time: 3.246826171875ms
init-time: 1.996826171875ms
init-time: 2.52099609375ms
init-time: 2.19287109375ms
init-time: 2.07080078125ms
```

**After**

```
init-time: 2.282958984375ms
init-time: 2.01708984375ms
init-time: 2.58203125ms
init-time: 1.857177734375ms
init-time: 2.635009765625ms
init-time: 2.766845703125ms
init-time: 1.570068359375ms
init-time: 2.3359375ms
init-time: 2.297119140625ms
init-time: 1.691162109375ms
init-time: 1.953857421875ms
init-time: 1.69384765625ms
init-time: 1.6259765625ms
init-time: 1.794921875ms
init-time: 1.988037109375ms
```

Memory allocation profile is same before/after during the initialization (around 1 MB) for both of the bundles