@vigneshshanmugam 2019-04-02
Will be taken over by @jahtalab 
Note

+ Service factory would still exist with a minimal code to hold the instances at runtime instead of file level instances