@jahtalab 2019-04-02
I've created this issue: https://github.com/elastic/apm-agent-rum-js/pull/231 to do the required changes. we can close this one.